+++ lib/src/main/java/org/eclipse/tsp/java/client/experiment/ExperimentApi.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 import org.eclipse.tsp.java.client.shared.query.Query;
 import org.eclipse.tsp.java.client.shared.restclient.RestClient;
 import org.eclipse.tsp.java.client.shared.tspclient.TspClientResponse;
-import org.eclipse.tsp.java.client.tsp.AbstractTspApi;
+import org.eclipse.tsp.java.client.tspapi.AbstractTspApi;
 
 public class ExperimentApi extends AbstractTspApi {
     final String EXPERIMENT_API_URL = "%s/experiments";
