+++ lib/src/main/java/org/eclipse/tsp/java/client/trace/TraceApi.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 import org.eclipse.tsp.java.client.shared.query.Query;
 import org.eclipse.tsp.java.client.shared.restclient.RestClient;
 import org.eclipse.tsp.java.client.shared.tspclient.TspClientResponse;
-import org.eclipse.tsp.java.client.tsp.AbstractTspApi;
+import org.eclipse.tsp.java.client.tspapi.AbstractTspApi;
 
 public class TraceApi extends AbstractTspApi {
     private final String TRACE_API_URL = "%s/traces";
